AdministratorPermissions moved to Common project
