+++ Api/Services/Management/ExternalAdminContext.cs
@@ -1,5 +1,5 @@
 using HappyTravel.Edo.Api.Infrastructure;
-using HappyTravel.Edo.Api.Models.Management.Enums;
+using HappyTravel.Edo.Common.Enums.Administrators;
 
 namespace HappyTravel.Edo.Api.Services.Management
 {
