+++ Api/Services/Management/HttpBasedAdministratorContext.cs
@@ -1,9 +1,7 @@
 using System.Threading.Tasks;
 using CSharpFunctionalExtensions;
 using HappyTravel.Edo.Api.Infrastructure;
-using HappyTravel.Edo.Api.Models.Management.Enums;
-using HappyTravel.Edo.Api.Models.Users;
-using HappyTravel.Edo.Common.Enums;
+using HappyTravel.Edo.Common.Enums.Administrators;
 using HappyTravel.Edo.Data;
 using HappyTravel.Edo.Data.Management;
 using Microsoft.EntityFrameworkCore;
